0x0C-python-almost_a_circle/models/rectangle.py
#!/usr/bin/python3
"""
Module containing class Rectangle
"""
from models.base import Base


class Rectangle(Base):
    """Rectangle
    Class rectangle, inherits from class Base

    Args:
        Base (class): Parent class

    Returns:
        _type_: _description_
    """
    @property
    def width(self):
        return self.__width

    @property
    def height(self):
        return self.__height

    @property
    def x(self):
        return self.__x

    @property
    def y(self):
        return self.__y

    @width.setter
    def width(self, width):
        self.__width = width

    @height.setter
    def height(self, height):
        self.__height = height

    @x.setter
    def x(self, x):
        self.__x = x

    @y.setter
    def y(self, y):
        self.__y = y

    def __init__(self, width, height, x=0, y=0, id=None):
        super().__init__(id)
        self.width = width
        self.height = height
        self.x = x
        self.y = y
